Refactor Book Service

+++ src/main/java/org/libraryaccountingproject/repositories/BooksRepository.java
@@ -7,7 +7,7 @@ import org.libraryaccountingproject.entities.Book;
 import java.util.List;
 
 @Repository
-public interface BooksRepository extends JpaRepository<Book, Long> {
+public interface BooksRepository extends JpaRepository<Book, Integer> {
 
     boolean existsById(Long id);
